# optimize eof jumpdest analysis