--- a/core/vm/analysis_legacy_test.go
+++ b/core/vm/analysis_legacy_test.go
@@ -130,4 +130,6 @@ func BenchmarkJumpdestOpEOFAnalysis(bench *testing.B) {
 	bench.Run(op.String(), bencher)
 	op = RJUMPV
 	bench.Run(op.String(), bencher)
+	op = EOFCREATE
+	bench.Run(op.String(), bencher)
 }